Move grid options parser into component
